requirements.txt
# --- Core Deep Learning ---
torch>=2.9.1
torchvision>=0.24.1
torchaudio>=2.9.1
transformers>=4.36.0

# --- Data Science & Visualization (Jupyter Essentials) ---
numpy>=1.25.0
pandas>=2.3.3
matplotlib>=3.10.8
seaborn>=0.13.2
tqdm>=4.67.1
ipykernel>=7.1.0
scikit-learn>=1.3.0

# --- Computer Vision & Image Augmentation ---
opencv-python>=4.8.1
albumentations>=1.3.1
pillow>=10.0.0

# --- Agentic AI Frameworks (Weeks 4-12) ---
langchain>=0.1.0
langgraph>=0.1.0
crewai>=0.1.0

pyautogen

# --- Deployment & Utilities ---
fastapi>=0.111.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
